ResponseTimer

+++ UptimeMonitor/Data/Entities/LogEntry.cs
@@ -7,6 +7,6 @@ public class LogEntry
     public Guid Id { get; set; }
     public string Url { get; set; }
     public HttpStatusCode StatusCode { get; set; }
-    public int ResponseTime { get; set; }
+    public long ResponseTime { get; set; }
     public DateTime Date { get; set; }
 }
